+++ lib/View/home_page.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatefulWidget {
+  HomePage({Key key, this.title}) : super(key: key);
+  final String title;
+
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+
+        title: Text(widget.title),
+      ),
+      body: Center(
+
+        child: Column(
+
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'App',
+            ),
+          ],
+        ),
+      ),// This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+}
